Partially restructured the project
