+++ amulet_map_editor/api/__init__.py
@@ -1,0 +1,2 @@
+from .version import version
+from .framework import AmuletUI
