+++ amulet_map_editor/api/datatypes.py
@@ -1,0 +1,17 @@
+from typing import Dict, Union, Callable, Tuple, Any
+
+MenuData = Dict[
+    str,
+    Dict[
+        str,
+        Dict[
+            str,
+            Union[
+                Callable,
+                Tuple[Callable],
+                Tuple[Callable, str],
+                Tuple[Callable, str, Any],
+            ],
+        ],
+    ],
+]
