+++ amulet_map_editor/api/framework/__init__.py
@@ -1,0 +1,1 @@
+from .amulet_ui import AmuletUI
