+++ amulet_map_editor/api/framework/programs/__init__.py
@@ -1,0 +1,2 @@
+from .base_program import BaseProgram
+from .about_program import AboutProgram
