+++ main.py
@@ -1,13 +1,12 @@
 import wx
-import amulet_map_editor.resources
-from amulet_map_editor.amulet_ui import AmuletMainWindow
+from amulet_map_editor import AmuletUI
 from amulet_map_editor import log
 import traceback
 
 if __name__ == "__main__":
     try:
         app = wx.App()
-        frame = AmuletMainWindow(None)
+        frame = AmuletUI(None)
         app.MainLoop()
     except Exception as e:
         log.critical(
